+++ stroom-core-server/src/main/java/stroom/proxy/repo/ProxyFileHandler.java
@@ -18,11 +18,7 @@ package stroom.proxy.repo;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import stroom.feed.MetaMap;
 import stroom.util.io.CloseableUtil;
-import stroom.util.io.FileUtil;
-import stroom.util.io.InitialByteArrayOutputStream;
-import stroom.util.io.InitialByteArrayOutputStream.BufferPos;
 import stroom.util.io.StreamProgressMonitor;
 import stroom.util.shared.ModelStringUtil;
 import stroom.util.thread.BufferFactory;
@@ -71,39 +67,39 @@ public final class ProxyFileHandler {
         return entrySequence;
     }
 
-    private long getRawContentSize(final Path file) throws IOException {
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("Getting raw content size for  '" + FileUtil.getCanonicalPath(file) + "'");
-        }
-
-        long totalSize = 0;
-
-        try (final StroomZipFile stroomZipFile = new StroomZipFile(file)) {
-            for (final String sourceName : stroomZipFile.getStroomZipNameSet().getBaseNameSet()) {
-                totalSize += getRawEntrySize(stroomZipFile, sourceName, StroomZipFileType.Meta);
-                totalSize += getRawEntrySize(stroomZipFile, sourceName, StroomZipFileType.Context);
-                totalSize += getRawEntrySize(stroomZipFile, sourceName, StroomZipFileType.Data);
-            }
-        }
-
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("Raw content size for  '" + FileUtil.getCanonicalPath(file) + "' is " + ModelStringUtil.formatIECByteSizeString(totalSize));
-        }
-
-        return totalSize;
-    }
-
-    private long getRawEntrySize(final StroomZipFile stroomZipFile,
-                                 final String sourceName,
-                                 final StroomZipFileType fileType)
-            throws IOException {
-        final long size = stroomZipFile.getSize(sourceName, fileType);
-        if (size == -1) {
-            throw new IOException("Unknown raw file size");
-        }
-
-        return size;
-    }
+//    private long getRawContentSize(final Path file) throws IOException {
+//        if (LOGGER.isDebugEnabled()) {
+//            LOGGER.debug("Getting raw content size for  '" + FileUtil.getCanonicalPath(file) + "'");
+//        }
+//
+//        long totalSize = 0;
+//
+//        try (final StroomZipFile stroomZipFile = new StroomZipFile(file)) {
+//            for (final String sourceName : stroomZipFile.getStroomZipNameSet().getBaseNameSet()) {
+//                totalSize += getRawEntrySize(stroomZipFile, sourceName, StroomZipFileType.Meta);
+//                totalSize += getRawEntrySize(stroomZipFile, sourceName, StroomZipFileType.Context);
+//                totalSize += getRawEntrySize(stroomZipFile, sourceName, StroomZipFileType.Data);
+//            }
+//        }
+//
+//        if (LOGGER.isDebugEnabled()) {
+//            LOGGER.debug("Raw content size for  '" + FileUtil.getCanonicalPath(file) + "' is " + ModelStringUtil.formatIECByteSizeString(totalSize));
+//        }
+//
+//        return totalSize;
+//    }
+//
+//    private long getRawEntrySize(final StroomZipFile stroomZipFile,
+//                                 final String sourceName,
+//                                 final StroomZipFileType fileType)
+//            throws IOException {
+//        final long size = stroomZipFile.getSize(sourceName, fileType);
+//        if (size == -1) {
+//            throw new IOException("Unknown raw file size");
+//        }
+//
+//        return size;
+//    }
 
     private void sendEntry(final List<? extends StroomStreamHandler> requestHandlerList, final StroomZipFile stroomZipFile,
                              final String sourceName, final StreamProgressMonitor streamProgress,
@@ -148,28 +144,28 @@ public final class ProxyFileHandler {
         }
     }
 
-    private void sendEntry(final List<? extends StroomStreamHandler> stroomStreamHandlerList, final MetaMap metaMap,
-                          final StroomZipEntry targetEntry) throws IOException {
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("sendEntry() - " + targetEntry);
-        }
-        final byte[] buffer = BufferFactory.create();
-        for (final StroomStreamHandler stroomStreamHandler : stroomStreamHandlerList) {
-            if (stroomStreamHandler instanceof StroomHeaderStreamHandler) {
-                ((StroomHeaderStreamHandler) stroomStreamHandler).handleHeader(metaMap);
-            }
-            stroomStreamHandler.handleEntryStart(targetEntry);
-        }
-        final InitialByteArrayOutputStream initialByteArrayOutputStream = new InitialByteArrayOutputStream(buffer);
-        metaMap.write(initialByteArrayOutputStream, false);
-        final BufferPos bufferPos = initialByteArrayOutputStream.getBufferPos();
-        for (final StroomStreamHandler stroomStreamHandler : stroomStreamHandlerList) {
-            stroomStreamHandler.handleEntryData(bufferPos.getBuffer(), 0, bufferPos.getBufferPos());
-        }
-        for (final StroomStreamHandler stroomStreamHandler : stroomStreamHandlerList) {
-            stroomStreamHandler.handleEntryEnd();
-        }
-    }
+//    private void sendEntry(final List<? extends StroomStreamHandler> stroomStreamHandlerList, final MetaMap metaMap,
+//                          final StroomZipEntry targetEntry) throws IOException {
+//        if (LOGGER.isDebugEnabled()) {
+//            LOGGER.debug("sendEntry() - " + targetEntry);
+//        }
+//        final byte[] buffer = BufferFactory.create();
+//        for (final StroomStreamHandler stroomStreamHandler : stroomStreamHandlerList) {
+//            if (stroomStreamHandler instanceof StroomHeaderStreamHandler) {
+//                ((StroomHeaderStreamHandler) stroomStreamHandler).handleHeader(metaMap);
+//            }
+//            stroomStreamHandler.handleEntryStart(targetEntry);
+//        }
+//        final InitialByteArrayOutputStream initialByteArrayOutputStream = new InitialByteArrayOutputStream(buffer);
+//        metaMap.write(initialByteArrayOutputStream, false);
+//        final BufferPos bufferPos = initialByteArrayOutputStream.getBufferPos();
+//        for (final StroomStreamHandler stroomStreamHandler : stroomStreamHandlerList) {
+//            stroomStreamHandler.handleEntryData(bufferPos.getBuffer(), 0, bufferPos.getBufferPos());
+//        }
+//        for (final StroomStreamHandler stroomStreamHandler : stroomStreamHandlerList) {
+//            stroomStreamHandler.handleEntryEnd();
+//        }
+//    }
 
     public void deleteFiles(final StroomZipRepository stroomZipRepository, final List<Path> fileList) {
         for (final Path file : fileList) {