+++ stroom-core-server/src/main/java/stroom/streamtask/server/FilePackProcessor.java
@@ -1,0 +1,7 @@
+package stroom.streamtask.server;
+
+import stroom.proxy.repo.StroomZipRepository;
+
+public interface FilePackProcessor {
+    void process(StroomZipRepository stroomZipRepository, FilePack filePack);
+}